Move options panel to form package
